pyproject.toml
[project]
name = "linkedin-posting-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "grandalf>=0.8",
    "langchain[mistralai]>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "uv>=0.7.12",
    "uvicorn>=0.34.3",
]
